Move tests for max idv attempts to shared example
**Why**: We want to be able to test that code to limit the number of idv
attempts works for both the OIDC and SAML flows.

Note that the specs that require Javascript are included in a new
feature spec file, since the code to send a request from a SP does not
work with JS enabled.
